pyproject.toml
[tool.poetry]
name = "giveaway bot"
version = "0.1.0"
description = "A simple giveaway bot"
authors = ["timelessnesses <mooping3roblox@gmail.com>"]
license = "MPL 2.0"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
orjson = "^3.6.8"
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
asyncpg = "^0.25.0"
uvloop = {version = "^0.16.0", platform = "linux"}
python-dateutil = "^2.8.2"
parsedatetime = "^2.6"
python-dotenv = "^0.20.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
jishaku = "^2.5.0"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
